Move ICraftingWatcherNode stuff to abstract superclass (see relevant code review and comments)
