+++ src/main/java/appeng/parts/reporting/ConversionMonitorPart.java
@@ -27,12 +27,9 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.phys.Vec3;
 
-import appeng.api.networking.IStackWatcher;
-import appeng.api.networking.crafting.ICraftingWatcherNode;
 import appeng.api.parts.IPartItem;
 import appeng.api.parts.IPartModel;
 import appeng.api.stacks.AEItemKey;
-import appeng.api.stacks.AEKey;
 import appeng.api.storage.ITerminalHost;
 import appeng.api.storage.MEStorage;
 import appeng.api.storage.StorageHelper;
@@ -71,27 +68,8 @@ public class ConversionMonitorPart extends AbstractMonitorPart implements ITermi
     public static final IPartModel MODELS_LOCKED_HAS_CHANNEL = new PartModel(MODEL_BASE, MODEL_LOCKED_ON,
             MODEL_STATUS_HAS_CHANNEL);
 
-    private IStackWatcher craftingWatcher;
-
     public ConversionMonitorPart(IPartItem<?> partItem) {
         super(partItem, true);
-
-        getMainNode().addService(ICraftingWatcherNode.class, new ICraftingWatcherNode() {
-            @Override
-            public void updateWatcher(IStackWatcher newWatcher) {
-                craftingWatcher = newWatcher;
-                configureWatchers();
-            }
-
-            @Override
-            public void onRequestChange(AEKey what) {
-            }
-
-            @Override
-            public void onCraftableChange(AEKey what) {
-                getMainNode().ifPresent(ConversionMonitorPart.this::updateReportingValue);
-            }
-        });
     }
 
     @Override
@@ -232,23 +210,6 @@ public class ConversionMonitorPart extends AbstractMonitorPart implements ITermi
                 player.containerMenu.broadcastChanges();
             }
         });
-    }
-
-    @Override
-    protected void configureWatchers() {
-        // Determine whether our configured item is craftable for requesting crafting jobs via the monitor
-        if (craftingWatcher != null) {
-            craftingWatcher.reset();
-        }
-
-        if (getDisplayed() != null) {
-            if (craftingWatcher != null) {
-                craftingWatcher.add(getDisplayed());
-            }
-        }
-
-        // Update base storage watcher and reported item amount via super call
-        super.configureWatchers();
     }
 
     @Override
